+++ src/numerical_analysis/discretization/__init__.py
@@ -1,5 +1,6 @@
 from . import mesh1d
 from .mesh1d import LineMesh
+from .mesh2d import PolygonMesh
 from .condition import Condition
 
-__all__ = ["mesh1d", "Condition", "LineMesh"]
+__all__ = ["mesh1d", "Condition", "LineMesh", "PolygonMesh"]