+++ src/numerical_analysis/fem/__init__.py
@@ -1,3 +1,4 @@
 from .fem1d import Fem1d
+from .fem2d import Fem2d
 
-__all__ = ["Fem1d"]
+__all__ = ["Fem1d", "Fem2d"]