+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.0.2 (2025/12/18)
+
+- Implemented code to numerically solve two-dimensional boundary-value problems using the FEM.
+    - `numerical_analysis.fem.Fem2d`, `numerical_analysis.discretization.PolygonMesh`, etc.
+
 ## 0.0.1 (2025/10/26)
 
 - The repository has been initialized.